add the structure of stack api
